pyproject.toml
[project]
name = "stabilization"
version = "0.1.0"
description = "A short video stabilization software for SkyPulse interview"
authors = [
    {name = "Tomer Buber",email = "tombub@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy (>=2.2.6,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.11.13"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["E501"]



